requirements.txt
requests>=2.31.0
urllib3>=2.0.0
cryptography>=41.0.0
PySocks>=1.7.1 